Mostly changes from onClick to handleClick Separated a couple of classes to their own files and updated the tested code into the article
