+++ 12-form-fields/b-textarea/src/App.js
@@ -1,12 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './App.css';
 
-class App extends Component {
-  render() {
-    return (
-      <textarea rows="4"/>
-    );
-  }
-}
-
-export default App;
+export default () => <textarea rows="4"/>;
