+++ 12-form-fields/h-radio/src/App.js
@@ -1,24 +1,6 @@
 import React, { Component } from 'react';
+import Radio from './Radio';
 
-class Radio extends Component {
- state = {
-   checked: !!this.props.checked
- }
- render () {
-   return <input
-     type="radio"
-     name={this.props.name}
-     value={this.props.value}
-     checked={this.state.checked}
-     onChange={e => this.onFieldChange(e)}
-   />
- }
- onFieldChange (e) {
-   this.setState({
-     checked: e.target.checked
-   });
- }
-}
 class App extends Component {
  render () {
    const colorRadioName = 'color';
