--- 12-form-fields/j-toggle/src/App.css
@@ -1,41 +1,0 @@
-.toggle-wrap {
-  position: relative;
-  display: inline-block;
-  width: 60px;
-  height: 34px;
-}
-.toggle-wrap input {display:none;}
-.toggle-el {
-  position: relative;
-  cursor: pointer;
-  border-radius: 34px;
-  height: 100%;
-  width: 100%;
-  background-color: #ccc;
-  -webkit-transition: .4s;
-  transition: .4s;
-}
-.toggle-el:before {
-  content: "";
-  position: absolute;
-  border-radius: 50%;
-  height: 26px;
-  width: 26px;
-  left: 4px;
-  bottom: 4px;
-  background-color: white;
-  -webkit-transition: .25s;
-  transition: .25s;
-}
-input:checked + .toggle-el {
-  background-color: #f32195;
-  box-shadow: inset 0 -20px 37px 0px rgba(0, 0, 0, 0.36);
-}
-input:focus + .toggle-el {
-  box-shadow: 0 0 1px #f32195;
-}
-input:checked + .toggle-el:before {
-  -webkit-transform: translateX(26px);
-  -ms-transform: translateX(26px);
-  transform: translateX(26px);
-}
