+++ 12-form-fields/j-toggle/src/App.js
@@ -1,28 +1,5 @@
 import React, { Component } from 'react';
-import './App.css';
-
-class Toggle extends Component {
- state = {
-   checked: !!this.props.checked
- }
- render () {
-   return (
-     <label className="toggle-wrap">
-       <div>{this.props.label}</div>
-       <input type="checkbox"
-         checked={this.state.checked}
-         onChange={this.onToggleChange}
-       />
-       <div className="toggle-el"></div>
-     </label>
-   );
- }
- onToggleChange = (e) => {
-   this.setState({
-     checked: e.target.checked
-   });
- }
-}
+import Toggle from './Toggle';
 
 class App extends Component {
  render () {
